+++ .gitignore
@@ -6,6 +6,7 @@
 *.exe
 *.o
 *.so
+telehash
 
 sb
